add getToken to hook folder
